+++ app/src/renderer/modules/principal-components-analysis/ComponentsCountPicker.tsx
