+++ app/src/renderer/modules/principal-components-analysis/DataVisualizer.tsx
