+++ app/src/renderer/modules/principal-components-analysis/LoadingsMatrix.tsx
