+++ app/src/renderer/modules/principal-components-analysis/Plot2D.tsx
