+++ app/src/renderer/modules/principal-components-analysis/PrincipalComponentsAnalysis.tsx
