+++ app/src/renderer/modules/principal-components-analysis/ScaleHandler.tsx
