+++ app/src/renderer/modules/principal-components-analysis/config/index.ts
