+++ app/src/renderer/modules/principal-components-analysis/config/reducer.ts
