+++ app/src/renderer/modules/principal-components-analysis/config/state.ts
