+++ app/src/renderer/modules/principal-components-analysis/config/steps.tsx
