+++ app/src/renderer/modules/principal-components-analysis/index.tsx
